chore(emails): reorganization of utils.py files
The code of these files were moved to the 'utils' module including those in the error message enum.
