+++ api_inmobiliaria/apps/emails/applications/account_management/__init__.py
@@ -1,1 +1,1 @@
-from .account_activation import AccountActivation, ActivationErrors
+from .account_activation import AccountActivation
