+++ api_inmobiliaria/apps/emails/infrastructure/views/account_management/send_token.py
@@ -1,7 +1,7 @@
 from apps.emails.infrastructure.db import TokenRepository
-from apps.emails.applications.managers import ActionLinkManager
-from apps.emails.utils import TokenGenerator
+from apps.emails.applications import ActionLinkManager
 from apps.users.infrastructure.db import UserRepository
+from apps.utils.generators import TokenGenerator
 from apps.utils.validators import is_valid_uuid
 from rest_framework.request import Request
 from rest_framework.response import Response
